Rename build to runner to match class name
This makes it more clear what you're including
